LNK-166 - Java DataRefinery Experiment
Initial implementation of the Java (Groovy) based message server aka Data refinery. An overview can be found in README.md.

At a high level, this project consumes JSON formatted messages, converts them to Neo4J cypher statements and applies them to the graph in batches. Batches are processed in multiple threads in an attempt to achieve high levels of throughput.

There is a series of tests that validate the correctness of the code. Some of the tests rely on external services like Neo4J and RabbitMQ. The tests can be run locally using testcontainers. See the readme for more information regarding that.