LNK 170 Introduced optional redis cache for caching connection nodes
This commit adds support for a redis cache for connection nodes. I've been getting a lot of contention with connection nodes as I scale up the server. I have a suspicion it's the connection merges that are getting hammered. This approach hashes each connection node parameter set and stores it in a redis cache. The code first checks the cache to see if the node has already been processed and if so, just ignores it.

This behaviour can be enabled with the USE_REDIS_CACHE environment variable.