LNK-170 Added redis docker container into RedisSpecification
This commit brings in another container into the testing infrastructure. Currently it is spun up from within th RedisSpecification when the RUN_WITH_TEST_CONTAINERS is set to Y. It's disconnected from the other containers (in it's own network), as it's currently not required that the other docker containers can see this one.