LNK-170 Added MessageServerCommand specification
Included improving support for shutting the server down from tests

Added totalMessagesProcessedByServer and totalErrorsProcessedByServer that keep track of the total number of messages and errors processed by the server. Useful for testing but I can also see that being interesting from a metrics point of view.

Fixed a bug with the RabbitMQ container not setting the port correctly on startup

Added a utility method to the Neo4j class to delete all the nodes created by tests

Added a two tests for the MessageServerCommand that populates the rabbit queue with a number of messages, starts the server, listens for messages and processes them.