LNK-213 Surface metrics for Prometheus
Initial support for surfacing metrics from the service. This implementation works in conjunction with micronaut prometheus implementation.

Timer metrics have been added to MessageBatchProcessor that records the time spent in the core processing methods.

I wasn't able to get Micronaut to inject a MeterRegistry into the MessageBatchProcessor class so I end up grabbing it from the environment and passing it in. Not ideal.

The other idea I've had that I can't get to work in micronaut is using some around advice via AOP to measure the methods in the service. I really don't like my current implementation of manually wrapping the process methods with timers, but it does give me the information that I want so I'll leave it for now. Something I would like to get back to.