LNK-231 Improved the Neo4J connection process.
Added support for the NEO4J_HOST property to optionally contain a protocol. Both bolt://localhost and localhost are now supported.

Added a specification to test the various Neo4j connection states

Added a connectionOk method to Neo4j and moved the logic for determining the version and edition to Neo4j class.