LNK-234 Added two new endpoints to the server for testing purposes.
GenerateTestMessagesController (/GenerateTestMessages) takes a numeric parameter and will generate that number of test messages through the service.

Employee messages are generate in a JSON format and published on the Rabbit queue.

DeleteTestMessagesController(/DeleteTestMessages) will remove all message from the Neo4J database that have a property of SourceSystem with a value of 'menome_test_framework'.