LNK-241 Rework Neo4J Connection Implementation
This was a fairly major rework of the way that connections and relationships are handled in the message server. The original approach wasn't handling the case where a node type could be found as a 'primary node' and a connection node in the same batch. In some cases the code was creating duplicate nodes with missing connections.

I decided to introduce to some concrete objects that represent Neo4JNodes and Neo4JConnections. These objects are responsible for generating the correct Cypher syntax given their attributes. I think this is clearer.

Batch processing is simpler now with this approach. All nodes are collected up, regardless if they are primary or connection nodes. They are grouped into microbatches by node type and graph merges are generated for each microbatch. Once the nodes have been merged into the graph, the connections are merged.